requirements.txt
fastapi==0.104.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tortoise-orm==0.20.0
# asyncpg==0.29.0
uvicorn[standard]==0.25.0
gunicorn==21.2.0